add utest rotation blinky

+++ blinky/source/movement/rotation.h
@@ -2,9 +2,9 @@
 #include "inttypes.h"
 
 typedef struct angle_degree_type {
-    uint8_t x;
-    uint8_t y;
-    uint8_t z;
+    int16_t x;
+    int16_t y;
+    int16_t z;
 } angle_degree;
 
 angle_degree calculate_moved_angle(void (*gyro_xyz)(float*),
